+++ src/main/kotlin/dev/m2en/citation/Main.kt
@@ -19,6 +19,7 @@ import dev.m2en.citation.command.*
 import dev.m2en.citation.command.chat.HelpCommand
 import dev.m2en.citation.command.message.onDebugMessageCommand
 import dev.m2en.citation.command.message.onRegister
+import dev.m2en.citation.handler.InteractionCommandInterface
 import dev.m2en.citation.message.QuoteDeleteListener
 import dev.m2en.citation.message.QuoteSendListener
 import io.github.cdimascio.dotenv.dotenv
