+++ src/main/kotlin/dev/m2en/citation/handler/InteractionCommandInterface.kt
@@ -1,4 +1,4 @@
-package dev.m2en.citation.command
+package dev.m2en.citation.handler
 
 import dev.kord.core.behavior.interaction.response.DeferredEphemeralMessageInteractionResponseBehavior
 import dev.kord.core.entity.interaction.GuildChatInputCommandInteraction
